pyproject.toml
[tool.poetry]
name = "kittytype-backend"
version = "0.1.0"
description = ""
authors = ["sprittzer <tania.zvr@yandex.ru>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
tortoise-orm = "^0.25.1"
asyncpg = "^0.30.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
pydantic-settings = "^2.10.1"
sendgrid = "^6.12.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.4.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"